everything that needed to be moved has been moved. Now I need to make sure everything works
