--- src/app/components/create.tweet.component.ts
@@ -1,27 +1,0 @@
-import {Component} from "@angular/core";
-import {TweetService} from "../services/tweet.service";
-import {Status} from "../shared/classes/status";
-import{Tweet} from "../shared/classes/tweet";
-
-
-
-@Component({
-	selector: "create-tweet",
-	templateUrl: "./templates/create-tweet.html"
-})
-
-export class CreateTweetComponent {
-
-	//declare needed state variables for later use.
-	status: Status = null;
-
-	tweet: Tweet = new Tweet(null,null,null,null);
-
-	constructor(private tweetService : TweetService) {}
-
-
-	createTweet(): void  {
-		this.tweetService.createTweet(this.tweet)
-			.subscribe(status => this.status = status);
-	}
-}
